% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pkg_and_install.R
\name{InstallPackagesIfNotAlready}
\alias{InstallPackagesIfNotAlready}
\title{Install Packages with Checks}
\usage{
InstallPackagesIfNotAlready(pkgs, install_if_not = TRUE,
  return_status = FALSE, verbose = FALSE)
}
\arguments{
\item{pkgs}{(character vector) names of packages to check installation status.}

\item{install_if_not}{(logical) whether to install missing packages or not.}

\item{return_status}{(logical) whether to return package installation status or not.}

\item{verbose}{(logical) whether to print package installation status or not.}
}
\value{
if \code{install_if_not} is \code{TRUE}, then return a \code{data.frame}
consisting of the installation status. The \code{data.frame} has two columns:
\code{"Before"} and \code{"After"} indicating the status before and after the
function \code{InstallPackagesIfNotAlready} is executed. Further, its rows are
named after the packages names as specified in \code{pkgs}.
}
\description{
Check Package Installation Status and Install If Missing Upon Request
}
\examples{
\dontrun{
InstallPackagesIfNotAlready(
  pkgs = c("MASS", "knitr"), 
  install_if_not = TRUE, 
  return_status = FALSE, 
  verbose = FALSE
)
}
}
\author{
Haoen Cui
}
